Debugging gyro (again)
Set the gyro back to the Rotate() approach, sinde the rotation=quaternion resulted in errors for specific values. Also finally fixed the bug with Rotate(), so that you set the rotation of the camera to the gyro.attitude at the start. (Bugfix: enabled the gyro earlier, since the data is not available from the start)